+++ NEWS.md
@@ -1,3 +1,14 @@
+# lmtp 1.5.3
+
+### New Features
+
+### Bug Fixes
+
+- Fixes bug where the `shifted` argument wouldn't work with multivariate exposure (see issue \#175).
+- Fixes bug where `mtp = FALSE` wouldn't work with multivariate exposure. 
+
+### General
+
 # lmtp 1.5.2
 
 ### New Features